Add base smart component. Refactor base component file name
