+++ src/base/component.js
